Switch to Boost Python
