+++ src/angle-estimator-python.hh
@@ -1,0 +1,3 @@
+#include <sot/dynamic-pinocchio/angle-estimator.h>
+
+typedef boost::mpl::vector< dynamicgraph::sot::AngleEstimator > entities_t;
