+++ src/integrator-force-exact-python.hh
@@ -1,0 +1,3 @@
+#include <sot/dynamic-pinocchio/integrator-force-exact.h>
+
+typedef boost::mpl::vector< dynamicgraph::sot::IntegratorForceExact > entities_t;
