+++ src/integrator-force-python.hh
@@ -1,0 +1,3 @@
+#include <sot/dynamic-pinocchio/integrator-force.h>
+
+typedef boost::mpl::vector< dynamicgraph::sot::IntegratorForce > entities_t;
