+++ src/integrator-force-rk4-python.hh
@@ -1,0 +1,3 @@
+#include <sot/dynamic-pinocchio/integrator-force-rk4.h>
+
+typedef boost::mpl::vector< dynamicgraph::sot::IntegratorForceRK4 > entities_t;
