+++ src/mass-apparent-python.hh
@@ -1,0 +1,3 @@
+#include <sot/dynamic-pinocchio/mass-apparent.h>
+
+typedef boost::mpl::vector< dynamicgraph::sot::MassApparent > entities_t;
