+++ src/waist-attitude-from-sensor-python.hh
@@ -1,0 +1,3 @@
+#include <sot/dynamic-pinocchio/waist-attitude-from-sensor.h>
+
+typedef boost::mpl::vector< dynamicgraph::sot::WaistAttitudeFromSensor > entities_t;
