+++ src/zmp-from-forces-python.hh
@@ -1,0 +1,3 @@
+#include "zmp-from-forces.h"
+
+typedef boost::mpl::vector< sot::dynamic::ZmpFromForces > entities_t;
