+++ src/zmpreffromcom-python.hh
@@ -1,0 +1,3 @@
+#include <sot/dynamic-pinocchio/zmpreffromcom.h>
+
+typedef boost::mpl::vector<dynamicgraph::sot::ZmprefFromCom> entities_t;
